Refactor db init to be in factory
