--- persephone_api/__init__.py
@@ -1,7 +1,0 @@
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
-
-# import the models *after* the db object is defined
-from . import db_models
