+++ persephone_api/api_endpoints/audio.py
@@ -4,7 +4,7 @@ This deals with the API access for audio files uploading/downloading.
 """
 import flask_uploads
 
-from .. import db
+from ..extensions import db
 from ..db_models import Audio
 from ..upload_config import audio_files, uploads_url_base
 from ..serialization import AudioSchema
