+++ persephone_api/api_endpoints/model.py
@@ -3,7 +3,7 @@ API endpoints for /model
 This deals with the API access for model definitions and metadata
 """
 
-from .. import db
+from ..extensions import db
 from ..db_models import TranscriptionModel
 from ..serialization import TranscriptionModelSchema
