+++ persephone_api/api_endpoints/transcription.py
@@ -4,7 +4,7 @@ This deals with the API access for transcription files uploading/downloading.
 """
 import flask_uploads
 
-from .. import db
+from ..extensions import db
 from ..db_models import Transcription
 from ..upload_config import text_files, uploads_url_base
 from ..serialization import TranscriptionSchema
