+++ persephone_api/extensions.py
@@ -1,0 +1,4 @@
+"""Each flask extension is initialized in the app factory located in app.py."""
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
